moved get People logic from router and backend-processing to PeopleController
